test: migrate subscription coverage to bun
